+++ source/networksvc.h
@@ -2,6 +2,7 @@
 
 #include "cy_result.h"
 #include <stdbool.h>
+#include <string.h>
 
 void network_service_task(void *param);
-cy_rslt_t network_svc_send_data(const char *buffer, int length) ;
+cy_rslt_t network_svc_send_data(const char *buffer, size_t length) ;
